% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{basic_feature_elimination}
\alias{basic_feature_elimination}
\title{Recursive feature elimination (Backwards Feature Selection) incorporating resampling. Returns a class object with selcted features properties
http://topepo.github.io/caret/recursive-feature-elimination.html}
\usage{
basic_feature_elimination(countdata, conditions, subset_sizes, repeats, number,
  method, metric, balanced = FALSE, lower_limit = 5)
}
\arguments{
\item{countdata}{- count data}

\item{conditions}{- conditions to be used for classification}

\item{subset_sizes}{- subset_sizes for the recursive feature elimination eg: subsetSizes <- c(2, 4, 6, 8).}

\item{repeats}{- the cross-validation procedure is repeated n times (2.1 of Algorithm 2)}

\item{number}{- number of folds for resampling}

\item{method}{eg: cv or repeatedcv}

\item{metric}{- metric to use e.g "Accuracy", "RMSE".}

\item{balanced}{- True or False (For stratified sampling)}
}
\description{
Recursive feature elimination (Backwards Feature Selection) incorporating resampling. Returns a class object with selcted features properties
http://topepo.github.io/caret/recursive-feature-elimination.html
}
