% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{check.bioconductor.packages}
\alias{check.bioconductor.packages}
\title{Check for bioconductor packages}
\usage{
check.bioconductor.packages(pkgs)
}
\arguments{
\item{pkg}{- relevant packages}
}
\description{
Check for bioconductor packages
}
