% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{check.packages}
\alias{check.packages}
\title{Check to see if packages are installed. Install them if they are not, then load them into the R session.}
\usage{
check.packages(pkg)
}
\arguments{
\item{pkg}{- relevant packages for the project}
}
\description{
Check to see if packages are installed. Install them if they are not, then load them into the R session.
}
