% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{check_input_samples}
\alias{check_input_samples}
\title{Checks if the minimum sample number is reached and if the two samples are balanced.}
\usage{
check_input_samples(conditions, lower_limit = 5)
}
\arguments{
\item{conditions}{labels of conditions per sample}
}
\description{
Checks if the minimum sample number is reached and if the two samples are balanced.
}
