% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{cross_validation}
\alias{cross_validation}
\title{Random forest model cross validation with n folds and m repeats}
\usage{
cross_validation(fold, repeats, countdata, conditions, ntree = NULL)
}
\arguments{
\item{fold}{- number of folds for resamplin}

\item{repeats}{- the cross-validation procedure is repeated n times}

\item{countdata}{count data matrix}

\item{conditions}{labels of count data matrix}

\item{ntree}{number of trees, default 100,000 if sample size >100, set to 10,000}
}
\description{
Random forest model cross validation with n folds and m repeats
}
