% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{cv_error}
\alias{cv_error}
\title{Get cross validation error based on all models used by cross validation}
\usage{
cv_error(cv)
}
\arguments{
\item{cv}{cross validation object, including for all repeats predictions, conditions and feature importance values}
}
\description{
Get cross validation error based on all models used by cross validation
}
