% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{cv_feature_importance}
\alias{cv_feature_importance}
\title{Get feature importance of all models used by cross validation}
\usage{
cv_feature_importance(cv)
}
\arguments{
\item{cv}{cross validation object, including for all repeats predictions, conditions and feature importance values}
}
\description{
Get feature importance of all models used by cross validation
}
