% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{dds_heatmap_differential_gene_counts}
\alias{dds_heatmap_differential_gene_counts}
\title{Generates heatmap for top {n} differential expressed genes and outputs to JSON file
based on normalized count data}
\usage{
dds_heatmap_differential_gene_counts(dds, dds_result, subset, datasets, outdir,
  meta, name, sample_list)
}
\arguments{
\item{dds}{- dds object}

\item{dds_result}{- dds result under specific conditions}

\item{subset}{- number of genes required}

\item{datasets}{- datasets used for the analysis}

\item{outdir}{- output directory}

\item{meta}{- meta data information}

\item{name}{- file name}

\item{sample_list}{- list of samples coresponding to conditions}
}
\description{
Generates heatmap for top {n} differential expressed genes and outputs to JSON file
based on normalized count data
}
