% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{dds_heatmap_distance}
\alias{dds_heatmap_distance}
\title{Generates heatmap of the sample distances and outputs to JSON file}
\usage{
dds_heatmap_distance(dds, datasets, outdir, method = "raw")
}
\arguments{
\item{dds}{- dds object}

\item{datasets}{- datasets used for the analysis}

\item{outdir}{- output directory}

\item{method}{- name of data transformation method}
}
\description{
Generates heatmap of the sample distances and outputs to JSON file
}
