% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{dds_heatmap_rld_vst}
\alias{dds_heatmap_rld_vst}
\title{Generates heatmap for the top {n} genes from Regularized Log transform 
of the count data and outputs to JSON file}
\usage{
dds_heatmap_rld_vst(dds, subset, datasets, outdir, method = "raw")
}
\arguments{
\item{dds}{- dds object}

\item{subset}{- number of genes required}

\item{datasets}{- datasets used for the analysis}

\item{outdir}{- output directory}

\item{method}{- name of data transformation method}
}
\description{
Generates heatmap for the top {n} genes from Regularized Log transform 
of the count data and outputs to JSON file
}
