% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{dds_PCA}
\alias{dds_PCA}
\title{Generates PCA dataframe for the dds object, outputs PCA coordinates 
to JSON file.}
\usage{
dds_PCA(dds, dataset_name, outdir, sample_group, transformation_name,
  plot = F)
}
\arguments{
\item{dds}{- dds object}

\item{outdir}{- output directory}

\item{sample_group}{- condition key names}

\item{plot}{- make scatter plot}

\item{dataset}{- dataset names}

\item{transformation}{- name of transformation}
}
\description{
Generates PCA dataframe for the dds object, outputs PCA coordinates 
to JSON file.
}
