% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{dds_TSNE}
\alias{dds_TSNE}
\title{Generates TSNE dataframe for the dds object for the first 3 dimensions
Explantion for the choice of parameters: https://cran.r-project.org/web/packages/Rtsne/Rtsne.pdf}
\usage{
dds_TSNE(dds, dataset_name, outdir, sample_group, perplexity,
  transformation_name, plot = F)
}
\arguments{
\item{dds}{- dds object}

\item{outdir}{- output directory}

\item{sample_group}{- condition key names}

\item{perplexity}{- perplexity of TSNE analysis}

\item{plot}{- make scatter plot}

\item{dataset}{- dataset names}

\item{transformation}{- name of transformation}
}
\description{
Generates TSNE dataframe for the dds object for the first 3 dimensions
Explantion for the choice of parameters: https://cran.r-project.org/web/packages/Rtsne/Rtsne.pdf
}
