% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{de_comparisons}
\alias{de_comparisons}
\title{DE Comparisons for all feasible combinations. Uses the concept of degree of freedoms and compares only
conditions in the sample group with the least differences. (n-1)}
\usage{
de_comparisons(dds, datasets, outdir, sample_group, formula, overwrite = T,
  shrink = T)
}
\arguments{
\item{dds}{- dds object}

\item{datasets}{- datasets used for the analysis}

\item{outdir}{- output directory}

\item{conditions}{- conditions from sample group}
}
\description{
DE Comparisons for all feasible combinations. Uses the concept of degree of freedoms and compares only
conditions in the sample group with the least differences. (n-1)
}
