% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{de_result}
\alias{de_result}
\title{Create DDS object file}
\usage{
de_result(dataset_data)
}
\arguments{
\item{dataset_data}{including count matrix, conditions and formula}
}
\description{
Create DDS object file
}
