% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{deseq_normalization}
\alias{deseq_normalization}
\title{Normalaize the dataset with DESeq normalization methods}
\usage{
deseq_normalization(dataset_data)
}
\arguments{
\item{dataset_data}{including count matrix, conditions and formula}
}
\description{
Normalaize the dataset with DESeq normalization methods
}
