% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{feature_logscale}
\alias{feature_logscale}
\title{Generates log scale of aggregated accuracy data of recursive feature elimination}
\usage{
feature_logscale(opt_data, agg_data, rfProfile)
}
\arguments{
\item{opt_data}{- reduced/optimized feature list}

\item{agg_data}{- aggregated accuracy data of recursive feature elimination}

\item{rfProfile}{- random forest model/profile}
}
\description{
Generates log scale of aggregated accuracy data of recursive feature elimination
}
