% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{filter.check.count.input}
\alias{filter.check.count.input}
\title{Check if count data for sample exists}
\usage{
\method{filter}{check.count.input}(sub_samples)
}
\arguments{
\item{sub_samples}{metda data of selected samples}
}
\description{
Check if count data for sample exists
}
