% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{filter.meta.data}
\alias{filter.meta.data}
\title{Filter dataset based on selected column and filter string.
The function returns all samples of a dataset if one sample
fulfils the filter criteria}
\usage{
\method{filter}{meta.data}(meta_data, filter, column)
}
\arguments{
\item{meta_data}{meta data}

\item{filter}{string to grep in column}

\item{column}{column to select}
}
\value{
data frame of filtered samples
}
\description{
Filter dataset based on selected column and filter string.
The function returns all samples of a dataset if one sample
fulfils the filter criteria
}
