% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{generate_filename}
\alias{generate_filename}
\title{Generates filename based on the type of data generated}
\usage{
generate_filename(type, outdir, ext, name = "", hash = F)
}
\arguments{
\item{type}{- type pf data (eg. heatmap,comparisons etc.)}

\item{outdir}{- output directory of file}

\item{ext}{- file extension (e.g .json)}
}
\description{
Generates filename based on the type of data generated
}
