% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{generate_res_meta}
\alias{generate_res_meta}
\title{Generate meta information to be stored in JSON file}
\usage{
generate_res_meta(dtset, sample_con1, sample_con2, cond_1, cond_2, sample_group,
  formula, genome)
}
\arguments{
\item{dtset}{- results}

\item{cond_1}{- condtions 2}

\item{cond_2}{- dataset}

\item{smp}{- condtions 1}
}
\description{
Generate meta information to be stored in JSON file
}
