% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{get.conditons}
\alias{get.conditons}
\title{Extract all conditions out of meta data and return a dataframe
compatible with deSeq2. Check if}
\usage{
get.conditons(dataset_metadata, formula)
}
\arguments{
\item{formula}{formula}

\item{meta_data}{meta data table}
}
\description{
Extract all conditions out of meta data and return a dataframe
compatible with deSeq2. Check if
}
