% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{get.metadata}
\alias{get.metadata}
\title{Extract datasets out of the meat data flat file and
combine all features from column new.sample.group to a new condition column}
\usage{
get.metadata(dataset, meta_data)
}
\arguments{
\item{dataset}{vector of dataset numbers}

\item{meta_data}{meta data table}
}
\value{
dataframe of metadata including requested dataset
}
\description{
Extract datasets out of the meat data flat file and
combine all features from column new.sample.group to a new condition column
}
