% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{get.metadata.single}
\alias{get.metadata.single}
\title{Extract a single dataset out of the meat data flat file and
combine all features from coloumn new.sample.group to condition}
\usage{
get.metadata.single(sub_sample)
}
\arguments{
\item{dataset}{dataset number}

\item{meta_data}{meta data table}
}
\value{
dataframe of metadata including requested dataset
}
\description{
Extract a single dataset out of the meat data flat file and
combine all features from coloumn new.sample.group to condition
}
