% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{get.run.ids}
\alias{get.run.ids}
\title{Get sample and run ids}
\usage{
get.run.ids(dataset, meta_data)
}
\arguments{
\item{dataset}{dataset number}

\item{meta_data}{meta data table}
}
\description{
Get sample and run ids
}
