% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{get.sample.counts}
\alias{get.sample.counts}
\title{Get transcript counts of sample list}
\usage{
get.sample.counts(sample)
}
\arguments{
\item{sample}{sample meta data}
}
\description{
Get transcript counts of sample list
}
