% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{get.sample.ids}
\alias{get.sample.ids}
\title{Get sample ids}
\usage{
get.sample.ids(dataset, meta_data)
}
\arguments{
\item{dataset}{dataset number}

\item{meta_data}{meta data table}
}
\description{
Get sample ids
}
