% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{get_comparisons}
\alias{get_comparisons}
\title{Get Comparisons for all feasible combinations. Uses the concept of degree of freedoms and compares only
conditions in the sample group with the least differences. (n-1)}
\usage{
get_comparisons(conditions, sample_group)
}
\description{
Get Comparisons for all feasible combinations. Uses the concept of degree of freedoms and compares only
conditions in the sample group with the least differences. (n-1)
}
