% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_final_model}
\alias{get_final_model}
\title{Calculate the final random forest model based on the reduced/optimized feature set
The final model can be balanced or un-balanced}
\usage{
get_final_model(countdata, conditions, balanced = FALSE, ntree = NULL,
  lower_limit = 5)
}
\arguments{
\item{countdata}{count data matrix}

\item{conditions}{labels of count data matrix}

\item{balanced}{specify if the model will build based on balanced input samples (default: FALSE)}

\item{ntree}{number of trees to use for random forest model}

\item{lower_limit}{- minimum number of samples per class (default: 5)}

\item{ntree}{number of trees, default 100,000 if sample size >100, set to 10,000}
}
\description{
Calculate the final random forest model based on the reduced/optimized feature set
The final model can be balanced or un-balanced
}
