% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{get_healthy_position}
\alias{get_healthy_position}
\title{To always get the same "fold change direction" we have to set the healthy comparison as first condition
return position of comparison}
\usage{
get_healthy_position(comparison_with_healthy, condition_matrix, disease_pos)
}
\description{
To always get the same "fold change direction" we have to set the healthy comparison as first condition
return position of comparison
}
