% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_mtry}
\alias{get_mtry}
\title{Number of variables randomly sampled as candidates at each split.}
\usage{
get_mtry(nfeature)
}
\arguments{
\item{nfeature}{number of maximal features/variables (genes)}
}
\description{
Number of variables randomly sampled as candidates at each split.
}
