% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{getClusterGenes}
\alias{getClusterGenes}
\title{Retrieves gene list given cluster number}
\usage{
getClusterGenes(cluster_no, train, kclusters)
}
\arguments{
\item{cluster_no}{cluster number}

\item{train}{training dataset}

\item{kclusters}{training dataset}
}
\description{
Retrieves gene list given cluster number
}
