% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.r
\name{heatmap}
\alias{heatmap}
\title{Write heatmap to json and plot heatmap}
\usage{
heatmap(mat, datasets, name, outdir, method = "", description = "",
  meta = NA, plot = F)
}
\description{
Write heatmap to json and plot heatmap
}
