% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{intersection_count}
\alias{intersection_count}
\title{Count number of identical features in condition_1 and condition_2.
If one of the features is NA, count this feature as identical}
\usage{
intersection_count(condition_1, condition_2)
}
\arguments{
\item{condition_1}{list of features of condition 1}

\item{condition_2}{list of features of condition 2}
}
\description{
Count number of identical features in condition_1 and condition_2.
If one of the features is NA, count this feature as identical
}
