% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{kmeansclustering}
\alias{kmeansclustering}
\title{Kmeans clustering step}
\usage{
kmeansclustering(train, no_clusters)
}
\arguments{
\item{train}{training dataset}

\item{no_clusters}{number of clusters}
}
\description{
Kmeans clustering step
}
