% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{normalize_data}
\alias{normalize_data}
\title{Normalaize the dataset with selected normalization method}
\usage{
normalize_data(dataset_data, method = "rpm")
}
\arguments{
\item{dataset_data}{count data, meta data and feature length}

\item{method}{available methods: rpm, rpkm, deseq}
}
\value{
normalized data
}
\description{
Normalaize the dataset with selected normalization method
}
