% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{oob}
\alias{oob}
\title{Out-of-bag error(OOB) obtained for each independent tree that forms the forest}
\usage{
oob(rf_model)
}
\arguments{
\item{rf_model}{random forest model}
}
\description{
Out-of-bag error(OOB) obtained for each independent tree that forms the forest
}
