% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.r
\name{pca}
\alias{pca}
\title{Generates PCA dataframe, outputs PCA coordinates 
to JSON file or plot.}
\usage{
pca(mat, labels, dataset_name = "dataset", outdir = "./", sample_group,
  name = "default", json = T, plot = F)
}
\arguments{
\item{mat}{- matrix}

\item{labels}{- labels for pca points}

\item{outdir}{- output directory}

\item{sample_group}{- condition key names}

\item{name}{- name prefix for output file}

\item{json}{- create json output}

\item{plot}{- make scatter plot}

\item{dataset}{- dataset name}
}
\description{
Generates PCA dataframe, outputs PCA coordinates 
to JSON file or plot.
}
