% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{preProcessing}
\alias{preProcessing}
\title{Preprocessing step to change data into the right format and splits into training and test datasets}
\usage{
preProcessing(count_matrix, conditions, pct)
}
\arguments{
\item{count_matrix}{count matrix}

\item{conditions}{conditions}

\item{pct}{- percentage for test and training data}
}
\description{
Preprocessing step to change data into the right format and splits into training and test datasets
}
