% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{rce_svm}
\alias{rce_svm}
\title{Returns a new dataframe with selected genes}
\usage{
rce_svm(train, no_clusters, required_no_clusters)
}
\arguments{
\item{train}{training dataset}

\item{no_clusters}{number of clusters}

\item{required_no_clusters}{required number of clusters}
}
\description{
Returns a new dataframe with selected genes
}
