% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{recursive_feature_elimination}
\alias{recursive_feature_elimination}
\title{Recursive feature elimination (Backwards Feature Selection) incorporating resampling. Returns a class object with selcted features properties
http://topepo.github.io/caret/recursive-feature-elimination.html}
\usage{
recursive_feature_elimination(countdata, conditions, method = "repeatedcv",
  repeats = 10, number = 5, metric = "Accuracy", balanced = F,
  two_step = F)
}
\arguments{
\item{countdata}{- count data}

\item{conditions}{- conditions to be used for classification}

\item{method}{eg: cv or repeatedcv}

\item{repeats}{- the cross-validation procedure is repeated n times (2.1 of Algorithm 2)}

\item{number}{- number of folds for resampling}

\item{metric}{- metric to use e.g "Accuracy", "RMSE".}

\item{balanced}{- specify if the model will build based on balanced input samples (default: FALSE)}

\item{two_step}{- use a second rfe two improve feature size}

\item{subset_sizes}{- subset_sizes for the recursive feature elimination eg: subsetSizes <- c(2, 4, 6, 8).}
}
\description{
Recursive feature elimination (Backwards Feature Selection) incorporating resampling. Returns a class object with selcted features properties
http://topepo.github.io/caret/recursive-feature-elimination.html
}
