% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{replace.na}
\alias{replace.na}
\title{Replace all <NA> values with new factor value of selected condition}
\usage{
replace.na(coldata, condition, new_value)
}
\arguments{
\item{coldata}{dataframe with condition values}

\item{condition}{condition to select}

\item{new_value}{replace <NA> with new_value}
}
\value{
condition dataframe
}
\description{
Replace all <NA> values with new factor value of selected condition
}
