% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{roc}
\alias{roc}
\title{Receiver Operating Characteristic curve (ROC) and area under this curve (AUC)}
\usage{
roc(rf_model)
}
\arguments{
\item{rf_model}{random forest model}
}
\description{
Receiver Operating Characteristic curve (ROC) and area under this curve (AUC)
}
