% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{rpm}
\alias{rpm}
\title{Normalaize the dataset with RPM or RPKM normalization method}
\usage{
rpm(dataset_data, rpkm = F)
}
\arguments{
\item{dataset_data}{count data, meta data and feature length}

\item{rpkm}{if true, returns rpkm}
}
\value{
rpkm
}
\description{
Normalaize the dataset with RPM or RPKM normalization method
}
