% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{set_number_of_trees}
\alias{set_number_of_trees}
\title{set number of trees based on sample size, due to performance the maximum tree number for
classification with more than 100 samples is set to 10,000 otherwise to 100,000}
\usage{
set_number_of_trees(ntree)
}
\arguments{
\item{ntree}{number of trees}
}
\description{
set number of trees based on sample size, due to performance the maximum tree number for
classification with more than 100 samples is set to 10,000 otherwise to 100,000
}
