% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{svm_scoring}
\alias{svm_scoring}
\title{SVM Scoring step ( runs svm on the clusters and scores clusters)}
\usage{
svm_scoring(train, kclusters)
}
\arguments{
\item{train}{training dataset}

\item{kclusters}{required number of clusters}
}
\description{
SVM Scoring step ( runs svm on the clusters and scores clusters)
}
