% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{train_model}
\alias{train_model}
\title{Train model with cross validation}
\usage{
train_model(count_data, conditions, model_type, repeats = 10, number = 5,
  balanced = FALSE, ntree = NULL)
}
\arguments{
\item{count_data}{count data, used for classification}

\item{conditions}{classification labels}

\item{model_type}{eg: 'rf', 'ranger'}

\item{repeats}{For repeated k-fold cross-validation only: the number of complete sets of folds to compute}

\item{number}{Either the number of folds or number of resampling iterations}

\item{balanced}{Balancing of imput samples}

\item{ntree}{number of trees, default 100,000 if sample size >100, set to 10,000}
}
\description{
Train model with cross validation
}
