% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.r
\name{tsne}
\alias{tsne}
\title{Generates TSNE dataframe, outputs TSNE coordinates 
to JSON file or plot.}
\usage{
tsne(mat, labels, perplexity = 5, dataset_name = "dataset", outdir = "./",
  sample_group, name = "default", json = T, plot = F)
}
\arguments{
\item{mat}{- matrix}

\item{labels}{- labels for pca points}

\item{perplexity}{- TSNE perpelexity}

\item{outdir}{- output directory}

\item{sample_group}{- condition key names}

\item{name}{- name prefix for output file}

\item{json}{- create json output}

\item{plot}{- make scatter plot}

\item{dataset}{- dataset name}
}
\description{
Generates TSNE dataframe, outputs TSNE coordinates 
to JSON file or plot.
}
