% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{write_package_list}
\alias{write_package_list}
\title{Write session Info to json file}
\usage{
write_package_list(file_name, dataset, version = 0.01)
}
\arguments{
\item{file_name}{output file name}

\item{dataset}{dataset name}

\item{version}{version number of toolbox}
}
\description{
Write session Info to json file
}
