% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{write_to_json}
\alias{write_to_json}
\title{Write data to json file}
\usage{
write_to_json(dataset, data, description, output_dir, file_name)
}
\arguments{
\item{dataset}{- data set name}

\item{data}{- arbitray data set}

\item{description}{- description of input data set}

\item{output_dir}{- path to output directory}

\item{file_name}{- json file name, without extension}
}
\description{
Write data to json file
}
