% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{aggregate_profiles}
\alias{aggregate_profiles}
\title{Aggregates the results from the random forest profile from recursive feature elimination step}
\usage{
aggregate_profiles(profiles)
}
\arguments{
\item{profiles}{- profiles from feature elimination of the dataset}
}
\description{
Aggregates the results from the random forest profile from recursive feature elimination step
}
