% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{feature_logscale}
\alias{feature_logscale}
\title{Generates log scale of the data
WIP - Work in Progress}
\usage{
feature_logscale(opt_data, agg_data, rfProfile, output_dir)
}
\arguments{
\item{opt_data}{Optimised data}

\item{agg_data}{- Aggregated data from rfprofile}
}
\description{
Generates log scale of the data
WIP - Work in Progress
}
