% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{get.metadata}
\alias{get.metadata}
\title{combine all features from coloumn new.sample.group to condition}
\usage{
get.metadata(sub_sample)
}
\arguments{
\item{dataset}{dataset number}

\item{meta_data}{meta data table}
}
\value{
dataframe of metadata including requested dataset
}
\description{
combine all features from coloumn new.sample.group to condition
}
